Add redis.js config
